redux email finished
