+++ src/components/PictName.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {Link} from "react-router-dom"
+import { Link } from "react-router-dom";
 import { MDBBox } from "mdbreact";
 
 import "../css/style.css";
@@ -21,7 +21,10 @@ const PictName = (props) => {
           <MDBBox>
             <img
               src={
-                props.bio.user_image === "https://firebasestorage.googleapis.com/v0/b/personal-email-d3b3b.appspot.com/o/images%2Fuser_image%2F9ff2f2f01c4bd1b013.png?alt=media&token=272b7417-a9fa-4fde-99b1-e00fa6d377ee" ? props.bio.user_image : require("../images/deafult-profile.jpg")
+                props.bio.user_image ===
+                "https://firebasestorage.googleapis.com/v0/b/personal-email-d3b3b.appspot.com/o/images%2Fuser_image%2F9ff2f2f01c4bd1b013.png?alt=media&token=272b7417-a9fa-4fde-99b1-e00fa6d377ee"
+                  ? props.bio.user_image
+                  : require("../images/deafult-profile.jpg")
               }
               alt="pict"
               style={{
@@ -36,44 +39,49 @@ const PictName = (props) => {
         </MDBBox>
         <p
           className="text-center mb-0 pt-2"
-          style={{ fontSize: "18px", fontWeight: "700", color: "#a13034"}}
+          style={{ fontSize: "18px", fontWeight: "700", color: "#a13034" }}
         >
           {props.bio.full_name}
         </p>
       </MDBBox>
-      <MDBBox className="text-center" style={{ margin: "0", padding: "0", marginTop:"10px" }}>
-      <Link to="/broadcast"> 
-        <button
-          color="transparent"
-          style={{
-            backgroundColor: "#f14c59",
-            border: "1px solid #f14c59",
-            color: "white",
-            boxShadow: "none",
-            borderRadius: "40px",
-            fontSize: "14px",
-            height: "40px",
-            margin: "0",
-            padding: "0 13px",
-          }}
-          className="text-capitalize"
-        >
-          <i className="fas fa-plus mr-1"></i> New Broadcast
-        </button>
+      <MDBBox
+        className="text-center"
+        style={{ margin: "0", padding: "0", marginTop: "10px" }}
+      >
+        <Link to="/broadcast">
+          <button
+            color="transparent"
+            style={{
+              backgroundColor: "#f14c59",
+              border: "1px solid #f14c59",
+              color: "white",
+              boxShadow: "none",
+              borderRadius: "40px",
+              fontSize: "14px",
+              height: "40px",
+              margin: "0",
+              padding: "0 13px",
+            }}
+            className="text-capitalize"
+          >
+            <i className="fas fa-plus mr-1"></i> New Broadcast
+          </button>
         </Link>
       </MDBBox>
       <hr />
       <MDBBox className="pl-3">
-        <p
+      <Link
+          to={props.active === "segments" ? "#" : "/segment-list"}
           className="mb-1"
           style={{
             fontSize: "18px",
             fontWeight: "600",
             color: "#192e35",
+            cursor : props.active !== "segments" ? "pointer" : "text"
           }}
         >
           All segments
-        </p>
+        </Link>
         {/* <a href="#!" style={{ color: "#192e35" }}>
           <p className="py-0 my-0" style={{ fontSize: "16px" }}>
             Woman
@@ -88,7 +96,8 @@ const PictName = (props) => {
       <hr />
       <MDBBox className="pl-3">
         <Link
-        to="/segment"
+          to="/segment"
+          onClick={props.params ? props.router : false}
           className="mb-1"
           style={{
             fontSize: "18px",
@@ -102,12 +111,28 @@ const PictName = (props) => {
       <hr />
       <MDBBox className="pl-3">
         <Link
-        to="/draft"
+          to={props.active !== "broadcasts" ? "/dashboard" : "#"}
           className="mb-1"
           style={{
             fontSize: "18px",
             fontWeight: "600",
             color: "#192e35",
+            cursor : props.active !== "broadcasts" ? "pointer" : "text"
+          }}
+        >
+          Broadcasts
+        </Link>
+      </MDBBox>
+      <hr />
+      <MDBBox className="pl-3">
+        <Link
+          to={props.active !== "draft" ? "/draft" : "#"}
+          className="mb-1"
+          style={{
+            fontSize: "18px",
+            fontWeight: "600",
+            color: "#192e35",
+            cursor : props.active !== "draft" ? "pointer" : "text"
           }}
         >
           Draft
@@ -116,12 +141,13 @@ const PictName = (props) => {
       <hr />
       <MDBBox className="pl-3">
         <Link
-        to="/database"
+          to={props.active !== "database" ? "/database" : "#"}
           className="mb-1"
           style={{
             fontSize: "18px",
             fontWeight: "600",
             color: "#192e35",
+            cursor : props.active !== "database" ? "pointer" : "text"
           }}
         >
           Database
