+++ src/routes/MainRoutes.jsx
@@ -15,8 +15,9 @@ import SignIn from "../pages/SignIn";
 import SignUp from "../pages/SignUp";
 import NotFound from "../pages/NotFound";
 import AboutUs from "../pages/AboutUs"
-import Accordion from "../pages/Accordion"
 import NewSegment from "../pages/NewSegment"
+import AllSegments from "../pages/AllSegments"
+import PreviewEmail from "../pages/PreviewEmail"
 const MainRoutes = () => {
   return (
     <Provider store={store}>
@@ -25,17 +26,18 @@ const MainRoutes = () => {
           <Route exact path="/" component={Home} />
           <Route exact path="/signin" component={SignIn} />
           <Route exact path="/signup" component={SignUp} />
-          <Route exact path="/broadcast" component={Accordion} />
-          <Route exact path="/broad" component={NewBroadcast} />
+          <Route exact path="/broadcast" component={NewBroadcast} />
+          <Route exact path="/broadcast/:mailId" component={PreviewEmail} />
           <Route exact path="/builder" component={EmailBuilder} />
           <Route exact path="/database" component={Database} />
           <Route exact path="/staff" component={DashboardStaff} />
           <Route exact path="/dashboard" component={Dashboard} />
           <Route exact path="/draft" component={DashboardDraft} />
-          <Route exact path="/draft/:draftId" component={Accordion} />
+          <Route exact path="/draft/:draftId" component={NewBroadcast} />
           <Route exact path="/about" component={AboutUs} />
-          {/* <Route exact path="/accord" component={Accordion} /> */}
           <Route exact path="/segment" component={NewSegment} />
+          <Route exact path="/segment/:segmentId" component={AllSegments} />
+          <Route exact path="/segment-list" component={AllSegments} />
           <Route component={NotFound} />
         </Switch>
       </Router>
